backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
fastapi-mail<2.0.0
python-jose
python-dotenv
google-auth
google-auth-oauthlib
alembic
bcrypt
pydub
opencv-python
dlib
numpy
